Split ev/interfaces/telegram_bot.py into a package by domain
The TelegramInterface class had grown to 2824 lines in a single file,
making it hard to navigate and maintain. Split it into
ev/interfaces/telegram_bot/ with one mixin module per domain (routing,
voice, media, pomodoro, keyboards, callbacks, commands_wrappers,
background_loops), keeping __init__/_post_init/_on_error/run() centralized
in base.py — same extract-and-recompose pattern used for memory.py,
commands.py and brain.py in the previous phases. No behavior changes.
