+++ ev/interfaces/telegram_bot/__init__.py
@@ -1,0 +1,3 @@
+from .base import TelegramInterface, run
+
+__all__ = ["TelegramInterface", "run"]
